refactor: move adding links logic to utility function
